Exercise 03.

+++ app/Http/Controllers/ArticleController.php
@@ -64,7 +64,12 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-        $article->delete();
-        return redirect()->route('articles.index')->with('success', 'Article deleted successfully');
+        if ($article->isDeletable())
+        {
+            $article->delete();
+            return redirect()->route('articles.index')->with('success', 'Article deleted successfully');            
+        }
+        else
+            return redirect()->route('articles.index')->with('error', 'Article stock must be 0 for deletion');
     }
 }
